Crate utils

Functions

get_home_dir
Gets the home directory of the current system. Will return correct path for windows, linux, and osx.

read_file_to_string
Read file as a String.

read_file_to_vec
Read file as a Vec<u8>

read_toml
Converts a toml file represented as a string to S

set_user_perm
Restricts permissions on a file to user-only: 0600

write_to_file
Writes a string to a specified file. Creates the desired path if it does not exist. Note: path and filename are appended to produce the resulting file path.