utils::cse
Function cse_fork

pub fn cse_fork(f: FuncMulti, point: &[f64], error: &[f64]) -> f64

Correctly rounds value of a function at the given point with proper error.

This uses the fork method to compute the error on the function. This method is only valid if the function is continuous and monotonic.

Arguments

f: A function that takes a slice of f64 and returns a f64. Each element of the slice represents a variable in the function.

point: A slice of f64 representing the point where the function is evaluated.

error: A slice of f64 representing the error on each variable of the function.