utils::derivation

Function richardson

pub fn richardson(
    f: FuncSingle,
    x: f64,
    h_init: f64,
    levels: usize,
    order: u8,
    accurate: bool,
) -> Result<Vec<Vec<f64>>, &'static str>

Richardson’s Extrapolation to approximate f’(x) at a particular x.

Arguments

f: Function to find derivative of.

x: Value of x to find derivative at.

h_init: Initial stepsize.

levels: Number of levels of extrapolation.

order: Order of the derivative to compute.

accurate: If true, use the accurate version of Richardson’s extrapolation, else use a more numerically stable version. See [Mode Comparison]

Returns

The Richardson table. The best estimate is at the bottom-right of the Richardson table: array[levels-1][levels-1].

Mode Comparison

Stable

This version increases h for better stability with h*2^n. The tradeoff is a worse approximation of the derivative as h is larger.

Pros

Avoids very small h -> less roundoff errors and subtractive cancellation errors.

Better stability in presence of noisy or tabulated data.

Cons

Requires more extrapolation steps to reach the same precision.

Might not achieve the same level of accuracy if h is too large initially.

Accurate

This version decreases h for better accuracy with h/2^n. The tradeoff is a worse stability of the derivative as h gets smaller and smaller.

Pros

Increased accuracy as h -> 0.

Enables higher-order convergence with fewer steps.

Cons

Floating-point roundoff error increases when h becomes too small.

Subtractive cancellation: f(x + h) - f(x - h) can suffer when h is tiny.