utils::error_propagation

Function cse_taylor

pub fn cse_taylor(f: FuncMulti, point: &[f64], error: &[f64]) -> f64

Correctly rounds value of a function at the given point with proper error.

Arguments

f: A function that takes a slice of f64 and returns a f64. Each element of the slice represents a variable in the function.

point: A slice of f64 representing the point where the function is evaluated.

error: A slice of f64 representing the error on each variable of the function.