utils

Module integration

Enums

QuadratureMethod

Functions

composite_quadrature

Computes the integral of a function using the Composite Quadrature Rule.

composite_quadrature_simpson13_error

composite_quadrature_simpson38_error

composite_quadrature_trapezoidal_error

Computes the error estimate for the Composite Trapezoidal Rule.

gauss_legendre_table

Gauss–Legendre quadrature nodes and weights for n = 1 to 5

gaussian_quadrature

Computes the integral of a function using Gaussian quadrature.

newton_cotes

Computes the integral of a Newton-Gregory Forward Polynomial Interpolation using the Newton-Cotes formula.

newton_cotes_error

romberg

Computes the integral of a function using the Romberg integration technique.