utils

Module interpolation

Enums

NonLinearRegressionType

SplineType

Functions

bezier

bezier_3d

bezier_mat

bspline

bspline_3d

bspline_mat

collocation_polynomial_surface

divided_difference

finite_diff_table

Computes the finite difference table for a given degree and function values.

lagrange

Computes the Lagrange interpolation polynomial at a given point x.

lagrange_basis_polynomial 🔒

Computes the Lagrange basis polynomial for a given index i.

lagrange_polynomial_error_range

Computes the error range of a Lagrange polynomial interpolation at a given point x.

linear_regression

Least Squares Interpolation

newton_divided_difference

newton_gregory_forward

Computes the Forward Newton-Gregory interpolation polynomial at a given point x.

newton_gregory_forward_error

nonlinear_regression

Least Squares Interpolation

spline

tridiagonal_solve