Function utils::io::connection_request

pub async fn connection_request(
    safe_window: &SafeWindow,
    connection_start: bool,
    srej: Option<u8>,
    tx: Sender<Vec<u8>>,
    condition: &SafeCond,
)

Sends a connection request frame to the server.

This function sends a connection request frame to the server and waits for the acknowledgment. It also handles resending the frame in case of a timeout.

The function returns when the connection is established.

Arguments

safe_window - The window to manage the frames.

connection_start - A boolean indicating if the connection is starting or ending.

srej - A boolean indicating if the connection uses REJ or SREJ.

tx - The sender channel to send the frame to the writer task.

condition - The condition variable to signal the window state.

Panics

The window lock fails

The frame fails to be pushed to the window

The frame fails to be sent to the writer task

The condition fails to wait