utils
Module iter

Structs

FilterDrainIter

Traits

DuplicateCheck

This trait provides a helper method to determine whether an Iterator contains any duplicates.

DuplicateCheckBy

This trait provides a helper method to determine whether an Iterator contains any duplicates using an accessor function.

FilterDrain

Helper trait that implements drain_filter which is not stable yet.

Functions

choose

Unzips a slice of pairs, returning items corresponding to choice

pick

Returns a subset of items in a collection which corresponds to provided indices

xor

XORs 2 slices of bytes of equal length