utils::iter
Trait DuplicateCheckBy

pub trait DuplicateCheckBy<'a, F, T, U>
where
    Self: Iterator<Item = &'a T>,
    F: Fn(&'a T) -> &'a U,
    T: 'a,
    U: 'a + Hash + Eq + Clone,
{
    // Provided method
    fn contains_dups_by(&mut self, f: F) -> bool { ... }
}

This trait provides a helper method to determine whether an Iterator contains any duplicates using an accessor function.

Provided Methods

fn contains_dups_by(&mut self, f: F) -> bool

Checks iterator for any duplicates using an accessor function