utils

Module linear_systems

Functions

create_augmented 🔒

Creates an augmented matrix by combining a matrix and a vector.

create_table_header 🔒

Creates and formats a header for the table.

create_table_row 🔒

Creates and formats a row for the table.

fixed_point_system

Solves non-linear systems using the fixed-point method.

full_piv_lu

Computes the LU decomposition of a matrix with full pivoting and returns the solution of the system.

gauss_seidel_solver

Performs the Gauss-Seidel algorithm to solve the system Ax = b.

is_diagonally_dominant

Checks if the matrix is diagonally dominant. This means that the absolute value of the diagonal element is greater than the sum of the absolute values of the other elements in the row.

iterative_solve

Solve a system of linear equations using the iterative method.

jacobi_solver

Performs the Jacobi algorithm to solve the system Ax = b.

ldu_decomposition

Decomposes a matrix A such that A = L + D + U

least_squares

Computes the solution of a linear system that minimizes the error on the solution.

lu_unpack

Unpacks the LU decomposition into the lower and upper triangular matrices.

make_triangular

Computes a triangular matrix with 1’s on the diagonal by applying row operations, including handling the augmented matrix (matrix + b), and returns the matrix and vector separately.

maximize_diagonal

Perform row permutations to make the diagonal elements the largest in each row.

newton_system

Solves non-linear systems using the Newton method.

no_piv_lu

Computes the LU decomposition of a matrix without pivoting.

no_piv_lu_packed

Computes the LU decomposition of a matrix without pivoting.

no_piv_lu_packed_mut

Computes the LU decomposition of a matrix without pivoting. The input matrix is modified in-place and contains the packed LU decomposition.

partial_piv_lu

Computes the LU decomposition of a matrix with partial pivoting and returns the solution of the system.

relaxation_solver

Performs the relaxation method to solve the system Ax = b.

solve_cramer

Computes the solution of a linear system using Cramer’s rule.

solve_gauss

Solves a linear system using the Gauss method.

solve_lu

Solves a linear system using an LU decomposition.

solve_triangular

putes the solution of a linear system using the triangle method.