utils
Module norms

Functions

condition_number

Compute the condition number of a matrix.

norm_mat

Compute the p-norm of a matrix.

norm_vec

Compute the p-norm of a vector.

x_error_bound_a

Computes the upper bound of the error on x given that the error on A is known.

x_error_range_b

Computes the range of the error on x given that the error on b is known.