utils::norms

Function x_error_range_b

pub fn x_error_range_b(
    a: &DMatrix<f64>,
    b: &DVector<f64>,
    b_star: &DVector<f64>,
    p: u32,
) -> Result<(f64, f64), &'static str>

Computes the range of the error on x given that the error on b is known.

Arguments

a: A matrix.

b: The true b vector.

b_star: The approximated b vector.

p: The norm to use. Use the 1 norm or the 0 (infinite) norm

Errors

If the matrix is singular.

If the vector norm is not compatible with the matrix norm.