utils::range
Struct RangeSet

pub struct RangeSet<T> { /* private fields */ }

A set of values represented using ranges.

A RangeSet is similar to any other kind of set, such as HashSet, with the difference being that the values in the set are represented using ranges rather than storing each value individually.

Invariants

RangeSet enforces the following invariants on the ranges it contains:

The ranges are sorted.

The ranges are non-adjacent.

The ranges are non-intersecting.

The ranges are non-empty.

This is enforced in the constructor, and guaranteed to hold after applying any operation on a range or set.

Examples

use utils::range::*;

let a = 10..20;

// Difference
assert_eq!(a.difference(&(15..25)), RangeSet::from([10..15]));
assert_eq!(a.difference(&(12..15)), RangeSet::from([(10..12), (15..20)]));

// Union
assert_eq!(a.union(&(15..25)), RangeSet::from([10..25]));
assert_eq!(a.union(&(0..0)), RangeSet::from([10..20]));

// Comparison
assert!(a.is_superset(&(15..18)));
assert!(a.is_subset(&(0..30)));
assert!(a.is_disjoint(&(0..10)));
assert_eq!(a.clone(), RangeSet::from(a));

Implementations

impl<T> RangeSet<T>

pub fn into_inner(self) -> Vec<Range<T>>

Returns the ranges of the set.

pub fn len_ranges(&self) -> usize

Returns the number of ranges in the set.

impl<T: Copy + Ord> RangeSet<T>

pub fn new(ranges: &[Range<T>]) -> Self
where Self: RangeUnion<Range<T>, Output = Self>,

Returns a new RangeSet from the given ranges.

The RangeSet is constructed by computing the union of the given ranges.

pub fn iter(&self) -> RangeSetIter<'_, T>

Returns an iterator over the values in the set.

pub fn iter_ranges(&self) -> RangeIter<'_, T>

Returns an iterator over the ranges in the set.

pub fn contains(&self, value: &T) -> bool

Returns true if the set contains the given value.

pub fn min(&self) -> Option<T>

Returns the minimum value in the set, or None if the set is empty.

pub fn end(&self) -> Option<T>

Returns the end of right-most range in the set, or None if the set is empty.

Note

This is the non-inclusive bound of the right-most range. See RangeSet::max for the maximum value in the set.

impl<T: Copy + Ord + Step + Sub<Output = T>> RangeSet<T>

pub fn max(&self) -> Option<T>

Returns the maximum value in the set, or None if the set is empty.

pub fn split_off(&mut self, at: &T) -> Self

Splits the set into two at the provided value.

Returns a new set containing all the existing elements >= at. After the call, the original set will be left containing the elements < at.

Panics

Panics if at is not in the set.

impl<T: Copy + Ord + Sub<Output = T>> RangeSet<T>

pub fn shift_left(&mut self, offset: &T)

Shifts every range in the set to the left by the provided offset.

Panics

Panics if the shift causes an underflow.

impl<T: Copy + Ord + Add<Output = T>> RangeSet<T>

pub fn shift_right(&mut self, offset: &T)

Shifts every range in the set to the right by the provided offset.

Panics

Panics if the the shift causes an overflow.

impl<T: Copy + Ord> RangeSet<T>
where Range<T>: ExactSizeIterator<Item = T>,

pub fn len(&self) -> usize

Returns the number of values in the set.

pub fn is_empty(&self) -> bool

Returns true if the set is empty.