Module utils::signal

Enums

Error

The error cases enumeration for signal handling.

Traits

Killable

Trait for threads that can be signalled via pthread_kill.

Functions

SIGRTMAX

Return the maximum (inclusive) real-time signal number.

SIGRTMIN

Return the minimum (inclusive) real-time signal number.

__libc_current_sigrtmax 🔒 ⚠

__libc_current_sigrtmin 🔒 ⚠

block_signal

Mask a given signal.

clear_signal

Clear a pending signal.

create_sigset

Create a sigset with given signals.

get_blocked_signals

Retrieve the signal mask that is blocked of the current thread.

register_signal_handler

Register the signal handler of signum.

sigrtmax

sigrtmin

unblock_signal

Unmask a given signal.

validate_signal_num

Verify that a signal number is valid.

Type Aliases

SignalHandler

Public alias for a signal handler. sigaction.

SignalResult

A simplified Result type for operations that can return Error.