Trait utils::sock_ctrl_msg::IntoIovec

pub unsafe trait IntoIovec {
    // Required methods
    fn as_ptr(&self) -> *const c_void;
    fn size(&self) -> usize;
}

Trait for types that can be converted into an iovec that can be referenced by a syscall for the lifetime of this object.

Safety

This is marked unsafe because the implementation must ensure that the returned pointer and size is valid and that the lifetime of the returned pointer is at least that of the trait object.

Required Methods

fn as_ptr(&self) -> *const c_void

Gets the base pointer of this iovec.

fn size(&self) -> usize

Gets the size in bytes of this iovec.