utils::tuple
Trait TupleTranspose

pub trait TupleTranspose {
    type Output;

    // Required method
    fn transpose(self) -> Option<Self::Output>;
}

A trait for transposing a tuple of options into an option of a tuple.

Required Associated Types

type Output

The output type of the transpose operation.

Required Methods

fn transpose(self) -> Option<Self::Output>

Transposes a tuple of options into an option of a tuple.