utils

Module utils

Enums

TableError

Defines the possible errors that can occur when formatting a table.

Constants

EPSILON

A f64 representing the perturbation used to compute the gradient.

Functions

choose_float

Computes the binomial coefficient “n choose k”.

compute_derivative

Computes the derivative of an arbitrary function at a given point.

compute_gradient

Computes the gradient of an arbitrary function at a given point.

factorial

Computes the factorial of a number.

function_vec

Computes at a given point the value of every function in a system of equations.

gradient_mat

Computes at a given point the gradient of every function in a system of equations.

table_formatter

This function formats a table with a header and contents.

Type Aliases

FuncMulti

This type defines a function that takes a slice of f64 and returns an f64. Each element of the slice represents a variable in the function.

FuncSingle

This type defines a function that takes a single f64 and returns an f64.