Crate utils

Modules

arg_parser

byte_order

epoll

Safe wrappers over the epoll API.

errno

Structures, helpers, and type definitions for working with errno.

eventfd

Structure and wrapper functions for working with eventfd.

fam

Trait and wrapper for working with C defined FAM structures.

ioctl

Macros and functions for working with ioctl.

kernel_version

net

Network-related utilities

rand

Miscellaneous functions related to getting (pseudo) random numbers and strings.

seek_hole

Traits and implementations over lseek64.

signal

sm

sock_ctrl_msg

Wrapper for sending and receiving messages with file descriptors on sockets that accept control messages (e.g. Unix domain sockets).

syscall

Wrapper for interpreting syscall exit codes.

tempdir

Structure for handling temporary directories.

tempfile

Struct for handling temporary files as well as any cleanup required.

terminal

Trait for working with termios.

time

validators

Macros

generate_fam_struct_impl

Generate FamStruct implementation for structs with flexible array member.

ioctl_ioc_nr

Declare a function that returns an ioctl number.

ioctl_iow_nr

Declare an ioctl that writes data.

skip_if_io_uring_supported

skip_if_io_uring_unsupported

Functions

get_page_size

Return the default page size of the platform, in bytes.

ioctl_expr

Function to calculate icotl number. Mimic of _IOC

u64_to_usize

Safely converts a u64 value to a usize value. This bypasses the Clippy lint check because we only support 64-bit platforms.

wrap_usize_to_u32

Converts a usize into a wrapping u32.