utils
Module utils

Enums

TableError

Defines the possible errors that can occur when formatting a table.

Constants

EPSILON

A f64 representing the perturbation used to compute the gradient.

Functions

compute_derivative

Computes the gradient of an arbitrary function at a given point.

compute_gradient

Computes the gradient of an arbitrary function at a given point.

function_vec

Computes the function value for the system of equations at a given point.

gradient_mat

Computes the gradient of the system of equations at a given point.

table_formatter

This function formats a table with a header and contents.

Type Aliases

FuncMulti

This type defines a function that takes a slice of f64 and returns an f64. Each element of the slice represents a variable in the function.

FuncSingle

This type defines a function that takes a single f64 and returns an f64.